requirements.txt
numpy
opencv-python==4.5.3.56
tensorboardX==2.4
Pillow
tqdm
thop
ptflops
torchstat 